Handle callback for redirect flow in Redirect class
